requirements.txt
discord.py
python-dotenv
aiohttp
cryptography
websocketspy-clob-client
